Clean TODOs in tests_wbemconnection.py and _wbemconnection.py
Fixes issue # 1240 - cvt test_wbemconnection.py TODOS to issues

Created a new issue #2327 that documents the TODOs that are still
current in this test module as 18 items to be worked on to complete the
tests of pywbem_mock.

This removes a number of obsolete TODOs and converts the others to
references to Issue #2327.  From there I can work them off.

Fixes one issue in _wbemconnection.py and clarifies one other TODO

1. Cleaned up code section that was marked with TODO about compiling
multiple files to actually do this correctly by building the search_path
for each file in the input files.

2. Remove one TODO since I confirmed that we are doing all the correct
tests