Issue #2062 part5b, refactor register_provider.
Refactored register_provider to clean up issues where we
were putting the register_provider in a method in the user-provider.

Goal is that in general the user should only have to:

1. Set up any defined prerequisites (ex. have an interop namespace for
the namespace provider)

2. Call register_provider with the defined inputs.

The register_provider should:

1. Confirm that prerequisites are satisfiled (normally the __init__)

2. Validate all of the parameters.

3. Register the provider

4. Call a post_register_setup so that the now active provider can do any
special setup (Ex. the namespace provider must be certain that there
is an instance of CIM_namespace for each existing namespace)

To do this, we defined a new method that can optionally go into
the provider post_register_setup(conn) that performs the provider
specific actions after the provider is registered modified
register_provider to attempt to call this method immediatly after
the provider is registered.

At this point, any actions taken by this method are executed on
the registered provider so, for example, the namespace provider
sets up all of the CIM_Namespace instances immediatly after
registration and before anything else takes place.

We also removed the code that was testing for an interop namespace
as part of the provider registration and simply made it a prerequisit
that the interop namespace exist before this provider is registered.

We may still have issues with conn in the future but for now, I defined
it as a requirement for the post_register_setup but, do not use it
for the __init__ since the only code is to find the interop_namespace
and that is a method of BaseProvider.

It is required for the post_register_setup so that this method could
issue client requests for anything that is required.

Note that if we get an exception back from post_register_setup we DO NOT
unregister the provider at this point. Since this is a mocker that
may be sufficient.