Regularize use of deepcopy for objects xfrd to and from repository
Since the .copy method for CIMClass and CIMInstance does not do a deep
copy and we want to insure that the objects in the dictionary are
truly independent copies of the ones in the input and outpu  we changed
to use deepcopy to copy all objects before they are placed into the
dictionary and for all objects that are extracted from the dictionary
with the copy property for both CIMInstance and CIMClass.

We were doing extraneous copies in _baserepository and so
removed these extra copy statements.

Adds tests for cim repository to assure that the objects are copied
on create, modify, get with copy and iter_value with copy.