issue # 2062 - part 4 update compiler interface to mocker
Updates the compiler interface to the compiler based on the
BaseRepositoryConnection to:

1. Implement the required methods as required for the mocker in the
mocker class _MockMOFWBEMConnection

2. Change the superclass for MocMOFWBEMConection to
BaseConnectionRepository

3. Require that the Faked_connection_object exist in construction.
   This class requires that there be a real repository since that
   is the goal of the compiler within the mocker.

4. Modify the create/modify/delete methods to get the data from
   the cim repository (i.e. self.conn)

4. Remove as much code as possible from the methods in
MOckMOFWBEMConnection  that duplicates code in the mock methods.
Note that in many cases this reduces the code to only call the
corresponding self.conn.... method.  The only method that really
uses significant code is CreateClass since that class handles
creating class dependencies and CreateInstance since we must have
at least all the key properties defined in any new instance to be
able to build the path for the cim repository.

5. Set the other methods to NotImplemented exception.

6. This pr also creates an implementation for ModifyClass since Modify
class is required in the CreateClass implementation in getting and
creating dependent classes in the repository. Since this is used
primarily in loading the schema where we encounter the same class
multiple times as a dependency, the alternate would have been to
drop the code that tries ModifyClass if the create fails with
ALREADY_EXISTS assuming that the new class is the duplicate of the
original class. I chose to implement ModifyClass.  NOTE: The reason
we did not see it before is that when I created the version earlier
where we were inserting directly into the repository, I simply did
an update if the create found an existing class.