Part of issue # 2062 - implement new schema compile method
The original method for compiling parts of a DMTF schema was
compile_dmtf_schema() which compiled a specific set of classes from the
DMTF schema defined by a DMTF schema version number and the directory
where a local version of that DMTF was supposed to be located.

This change separates the compile of classes of the schema from the process
of downloading and installing the schema.

Thus, originally, the FakedWBEMConnection method to create a mock
repository with some classes from the schema was:

conn.compile_dmtf_schema(DMTF_TEST_SCHEMA_VER, TESTSUITE_SCHEMA_DIR,
                         class_names=[CIM_Namespace,'CIM_ObjectManager'],
                         verbose=True)

the separation functionality yields:

schema = DMTFCIMSchema(DMTF_TEST_SCHEMA_VER, TESTSUITE_SCHEMA_DIR,
                       verbose=False)

conn.compile_dmtf_classes([CIM_Namespace,'CIM_ObjectManager']
                          schema.schema_mof_dir,
                          namespace=True)

In addition we changed the internal name in DMTFCIMSchema class from
schema_mof_file to schema_pragma_file.  Note that the property
schema_mof_file is retained but marked deprecated.  This helps to
clarify the meaning of this file.

Added tests for the new method to compile DMTF classes. Note that this
was a very small change, just extending the test class for the original
compile_dmtf_schema to execute both the original compile and the new
compile.

Added entry to changes.rst

Changes to change input parameters in compile_schema_mof

Modified compile_schema_mof to use the pragma file in place of the
mof_directory as input.

Changed schema_classes parameter on multiple methods to class_names.