Refactor register_provider
The original form for the register_provider and its relation to
namespace provider meant that the register_provider code was in the
namespace provider.

This refactors register_provider that so:

1. The provider can define a method post_register_setup(conn) which
register_provider will attempt to call.  Note that right now if that
method is not defined, the attribute error is just ignored. (DISCUSSION:
Require that each provider implement this or just ignore the exception)

This post_register_setup method is called with conn as a parameter
because it may be required to execute client operations as part of the
setup of the provider (It is required for the namespace provider to
create the namespace instances). Further, it does not cause any real
issues since this is used only by the method so should not create any
circular references for python.

2. The namespace provider is no longer responsible for creating the
interop namespace.  An interop namespace must exist before the provider
is created or the constructor will generate an exception. Simplifies the
provider.

3.  This removes any requirement to pass the name of the
interop_namespace to the provider since it already exists and there is a
method to find its name.

4. This means that any provider can elect to get the conn in its
constructor since the provider construction is outside the
register_provider but only those providers that need the conn during
normal execution of their user-defined methods need to either include
conn in the constructor parameters or save it in the provider.