Partly fixes #2062: add namespace provider (#2221)
* Part of issue # 2062 - implement new schema compile method

The original method for compiling parts of a DMTF schema was
compile_dmtf_schema() which compiled a specific set of classes from the
DMTF schema defined by a DMTF schema version number and the directory
where a local version of that DMTF was supposed to be located.

This change separates the compile of classes of the schema from the process
of downloading and installing the schema.

Thus, originally, the FakedWBEMConnection method to create a mock
repository with some classes from the schema was:

conn.compile_dmtf_schema(DMTF_TEST_SCHEMA_VER, TESTSUITE_SCHEMA_DIR,
                         class_names=[CIM_Namespace,'CIM_ObjectManager'],
                         verbose=True)

the separation functionality yields:

schema = DMTFCIMSchema(DMTF_TEST_SCHEMA_VER, TESTSUITE_SCHEMA_DIR,
                       verbose=False)

conn.compile_dmtf_classes(schema.schema_mof_dir,
         schema_classnames=[CIM_Namespace,'CIM_ObjectManager'],
         namespace=True)

In addition we changed the internal name in DMTFCIMSchema class from
schema_mof_file to schema_mof_pragma_file.  Note that the property
schema_mof_file is retained but marked deprecated.  This helps to
clarify the meaning of this file.

Added tests for the new method to compile DMTF classes. Note that this
was a very small change, just extending the test class for the original
compile_dmtf_schema to execute both the original compile and the new
compile.

Added entry to changes.rst

* Partly fixes issue # 2062, Adding namespace provider

This required changing components in FakedWBEMConnection and
DMTFCIMSchema classes because we changed some of the parameters
of provider registration.

It also adds a method to FakedWBEMConnection,
install_namespace_provider() that serves as a temporary method
to installe the namespace provider to be installed.

This pr adds three files:

1. pywbem_mock/config.py. Moved config variables from
FakedWBEMConnection, etc. to this common place.

2. _namespaceprovider, the first implementation of a user-defined
provider.

3. tests/unittest/pywbem_mock/testnamespaceprovider.py

4. Moved all of the provider registration methods to a new class
   ProviderRegistry.

5. Moved 2 methods from mainprovider to baseprovider to eliminate
   pylint issues.

6. Pulled the pywbem_mock configuration variables into a separate
   config.py to avoid importing back and forth between the
   other classes.

7. Moved the code for CIM_Namespace instances CreateInstance, etc.
   into the new provider

8. Modified the method compile_schema_classes to accept multiple
   pragmas so the user could supply multiple search paths and put
   user MOF into different directories than the DMTF schema classes

9. Removed old code that handled the InvokeMethod callback and
   registered callbacks.

10. Added methods to identify interop namespaces to WBEMConnection
   and BaseProvider.

11. Changed the name of the server in test_wbemserver.py and
   WBEMServerMock.py from  OpenPegasus which triggered the use
   of PG_Namespace to 'Mock_server' so that the class for
   namespace management is  CIM_Namespace.

12. Added a new test for the mocker (test_system_providers.py) which
   creates and uses the CIM_Namespace provider.

13. Modified test_wbemserver.py to use the new CIM_Namespace provider.

14. Added a method to the new namespace provider `install_provider`
    that handles the install of the provider.   Thus, the only
    code in FakedWBEMConnection for the install of the provider
    is creating the provider object and calling the providers
    install_provider() method. This is temporary and will be
    replaced by extending register_provider.

Remove several methods that were duplicated in both BaseProvider and
MainProvider.  They should have been only in BaseProvider.  This
includes the get_..._store methods and the add/remove namespace
methods.

* Corrections from comments Sat, 23 May