Fixes issue #2233, display of registered providers
1. Reorganized the internals of ProviderRegistry to make namespace the
top level.

2. Add display method. This displays each provider registration.

It also adds methods to get access to the registry by namespace,
classnames, or type and modifies the code to use these methods
in place of the internal structure of the repository for
the get_registered_provider method.  This displays information on the
currently registered providers.  The following is an example:

Registered Providers:
namespace: root/cimv2
  CIM_Foo      instance  UserInstanceTestProvider2
  CIM_Foo_Sub  instance  UserInstanceTestProvider2
namespace: root/cimv3\n'
  CIM_Foo      instance  UserInstanceTestProvider2
  CIM_Foo_Sub  instance  UserInstanceTestProvider2

3 Adds a new test of the display method with a variety
of inputs