Fixes issue # 2229, clean up convience functions
1. Remove the private version of the get_xxx_store functions in
FakedWBEMConnection and the public one in BaseProvider. All references
now use conn.cimrepository or cimrepository directly to get the
appropriate object store.

2. Make the instance attributes private with a public property for those
that should be public.  Currently the only public variable is
cimrepository although this meant that we had to modify many uses of
mainprovider in the tests.

3. Modify interop_namespace_names to become a property rather than
a method