Correct issues in Pywbem 1.0.0 found when testing with pywbemcli.
Correct an issue where we had function parameters misordered
and were not always returning the correct context.  This was
an issue introduced in pywbem 1.0.0 when we move everything
to MainProvider so there is no backword port necessary.

This also adds two tests since we were not testing the
OpenEnumerateInstances and OpenEnumerateInstancePaths requests
in test_wbemconnection.py.

Remove code that was inserting extra properties.
Removed code that I had somehow put into pywbem 1 that added properties
in the class but not in the instance to the instance.  No idea why I
thought that was a good idea.

Change the exception on InvokeMethod when method not found from
NOT_AVAILABLE to NOT_FOUND.  NOT_FOUND is higher in the list
of exception messages so it should be the correct status code.
This is pywbem v 1.0.0 only issue since I made the mistake of
modifying this status code when creating the MethodProvider.