Fix issue #2244 - Review Provider Manager
In response to the questions we made the following changes:

1. Move the definition of the default InstanceWriteProvider and
MethodProvider to the ProviderDispatcher. This is done with a property
for each that only constructs the default provider when required.

2. Remove the constructor for the default providers from _wbemconnection
so the only parameter for the construction of ProviderDispatcher is the
CIM repository.

3. Change the default MethodProvider behavior to actually call the
MethodProvider where that provider generates the required exception.
This maintains the same pattern as the other operations in the
ProviderDispatcher.

4. Change that exception to CIM_ERR_NOT_FOUNd