Fixes #2233: display registered providers (#2246)
* issue #2062 part5b, refactor register_provider.

TODO

* Refactor register_provider

The original form for the register_provider and its relation to
namespace provider meant that the register_provider code was in the
namespace provider.

This refactors register_provider that so:

1. The provider can define a method post_register_setup(conn) which
register_provider will attempt to call.  Note that right now if that
method is not defined, the attribute error is just ignored. (DISCUSSION:
Require that each provider implement this or just ignore the exception)

This post_register_setup method is called with conn as a parameter
because it may be required to execute client operations as part of the
setup of the provider (It is required for the namespace provider to
create the namespace instances). Further, it does not cause any real
issues since this is used only by the method so should not create any
circular references for python.

2. The namespace provider is no longer responsible for creating the
interop namespace.  An interop namespace must exist before the provider
is created or the constructor will generate an exception. Simplifies the
provider.

3.  This removes any requirement to pass the name of the
interop_namespace to the provider since it already exists and there is a
method to find its name.

4. This means that any provider can elect to get the conn in its
constructor since the provider construction is outside the
register_provider but only those providers that need the conn during
normal execution of their user-defined methods need to either include
conn in the constructor parameters or save it in the provider.

* Fixes issue #2233, display of registered providers

1. Reorganized the internals of ProviderRegistry to make namespace the
top level.

2. Add display method. This displays each provider registration.

It also adds methods to get access to the registry by namespace,
classnames, or type and modifies the code to use these methods
in place of the internal structure of the repository for
the get_registered_provider method.  This displays information on the
currently registered providers.  The following is an example:

Registered Providers:
namespace: root/cimv2
  CIM_Foo      instance  UserInstanceTestProvider2
  CIM_Foo_Sub  instance  UserInstanceTestProvider2
namespace: root/cimv3\n'
  CIM_Foo      instance  UserInstanceTestProvider2
  CIM_Foo_Sub  instance  UserInstanceTestProvider2

3 Adds a new test of the display method with a variety
of inputs

* Modify methods to return lists per comments

Modified provider_namespaces and provider_classnames to return lists
rather than the dict, refactored internal code to not use these methods,
and modified tests to test with case.  Note that the display... method
did not need modification since it only uses the methods in for loops.