Fixes issue #2247, MOF compiler changes default_namespace
The method compile_string changes the default_namespace in the
MOF_Compiler handle parameter which is typically the connection
object and if compiling into a namespace other than the default,
this namespace name becomes the connection default namespace.

The change saves and restores the default namespace within the
method compile_string.

However, most of the tests in mof_compiler and even a couple in the
mocker used this so those tests had to change also.