Expand MOF compiler to allow other namespaces
Extended the MOF compiler to avoid using default_namespace internally and
using compile_string() to set the namespace used by the compiler.

This required extending MOFWBEMConnection and MOCKMOFWBEMConnection to
add a namespace parameter to many of the method calls.

Adds simple test for namespace pragma in test_mof_compiler.py