fixes issue #2294 - Duplications in changes.rst incompatibilities.
NOTE: Since you took this issue, consider this what was the easiest way
document changes, etc. since it is formatted, complete and easy to
compare with original (sort of).  Feel free to use or just use as
input.

My proposed changes to changes.rst to remove some of the duplication and
extend explanations.

1. Shortened the summary section to one-liners essentially.

2. Added more explanation to the details section as to why,
alternativesm,  etc.