Make BaseProvider.get_bare_instance() private
The method gets an instance_store parameter, even though it is clear
that it always would use the instance store for the namespace of the
instance to be found, in the CIM repository that is known by the
BaseProvider object.

The reason for passing the instance store is that its lookup includes
a dictionary access with the namespace name as a key, and passing it
in allows doing that only once in case of loops. The only loop that
benefits from that so far is MainProvider.EnumerateInstances.

This change makes get_bare_instance() private and changes some of
its uses to use BaseProvider.object_exists() instead. That leaves
3 uses of the private method in MainProvider and InstanceWriteProvider,
in code that already uses other private methods of BaseProvider.

That allows defining a more properly defined get instance method in the
future.

This change also fixes the incorrect use of vertical bar in the
format string of an example, and simplifies the text of error messages
that show an instance path by removing the namespace because it is part
of the instance path already.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>