Fixes issue # 2310, update notedbook for pywbem_mock to reflect pywbem 1.0.0
Extended the notebook so that the deprecated/removed code for
compile_dmtf_schema is only conditionally used allowing this notebook to
be used with old versions of pywbem.

Corrects code to reflect the use of the method compile_dmtf_classes vs
compile_dmtf_schema and replaces the code that used the original
callback for method definition in the server with the user-defined
method provider.