Fixes #2277:  MOF compiler ignores EmbeddedObject instance property (#2331)
* Fixes issue #2277 - MOF compiler ignores EmbeddedObject instance property

The DMTF definition of the value of instance property MOF is one
or more strings containing the MOF that would be compiled into
CIM instances.

Originally the compiler simply passed the string through the compiler so
what went into a repository was simply a string for the value of the property.

The MOF compiler was extended to compile properties with the
EmbeddedInstance or EmbeddedObject qualifiers into CIMInstance
object(s).

The enhanced compiler creates CIMInstance objects so that the
instance compiled has a one or an  array of CIMInstance objects
corresponding

Limitations: The defintion of EmbeddedObject either an instance
or a class.  Since there are not instances of classes as
EmbeddedObject properties in the DMTF schema the compiler only
supports compiling instances.

Removed test for  embedded_object attribute in mocker providerdispatcher

* Changes per comments

1. Changed text in test_instance.mof

2. Added 2 more pieces of code to the compiler to test for the
embedded_object mode.

3. Fixed two places where we left in diagnostics (verbose=True) in
tests.