Add more tests for pywbem_mock
This pr adds a more detailed test of the internal _get_instance to test
for local_only, (calling from the request operations forces this to a
default) and extends other verifications in this test.

It also modifies the instance naming on the tst_instances and
tst_instances_mof fixture to produce the same instances so that either
can be used in tests with the same result.

NOTE: It corrects one issue in that we were deleting from the property
list without making the input a list.