Add validity test to pywbem_mock CreateClass
The CreateClass _mockmofwbemconnection used by the compiler .py for
existence of class dependencies (reference classes and embeddedobject
classes). However, the CreateClass in _mainprovider (direct user
call to create class does not validate dependent classes existence.

We could have pulled these together into a single function except that
there is a core difference between the two in that a direct user does
not have to maintain the local class cach used by other methods in
_mockmofwbemconnection.

Added validity tests to MainProvider.CreateClass for existence of
dependent classes.

At the same time we noted in the DMTF specification that it is allowed
for the EmbeddedObject value to be None which the code in the MOF
compiler does not do.  That test was added.

Finally, we noted that if the dependencies, reference class or
EmbeddedObject class were the same class as the target class the whole
process of trying to find the class was still executed so a test for
equality of the dependent class names with the class defined by
CreateClass was added.  Note that there was already a test for this
defined in the tst_classes_mof fixture.