Fixes issue # 2458, Assert from mof_compiler in dependency compile
The mof compiler failed with an assert if a class compile had
dependencies that were not in the repository or were invalid so the
dependency resolution failed.

Changed to use MOFDependencyError and also show the originating
exception since that may present more information about the failure.

Added tests for this issue to:
tests/unittest/pywbem_mock/test_wbemconnection.py

Note: We also modified one of the Exception messages in
pywbem_mock/_resolver to clarify it.