Fixes and tests for use of compiler namespace pragma This pr fixes issue #2256 as follows:
1. Adds tests to mof compiler for no name on the namespace pragma
just to be sure we have that covered

2. Adds tests in pywbem_mock for use of the namespace pragma.

3. Adds fix to the mof compiler for case where pybwbem_mock
compile_mof_string and compile_mof_file fail if the MOF contains a
namespace pragma that has a different namespace name than the namespace
parameter on the methods.  Note that it is just that one of the
MOFWBEMConnection lists does not get created for the namespace defined
in the pragma.

4. Extends documentation on the namespace parameter for the
compile_mof_string and compile_mof_file() methods to clarify the
behavior with the namespace pragma in the MOF.

5. Fix issue in mof compiler where a possible namespace path
with host could be passed to the namespace pragma but the compile
really did not handle the host component.  Now parses the
name to be sure it is a namespace name and no other namespace
components exist.

6. Adds the test to assure that the namespace name in the namespace pragma
is the namespace name and no other components of the namespace path.

7. Update documentation for namespace component for compile_mof_.
Update the documentation of the namespace parameter in pywbem_mock
compile_mof_string() and compile_mof_file() to reflect the requirenents
on the namespace parameter() to reflect the requirenents
on the namespace parameter.

8 Adds note to the list of compiler limitations in _mof_compiler.py
noting that the namespace pragma only accepts namespace name and not
namespace path as defined in the DMTF specification DSP0004.