Adds index section to generated documentation
This pr adds a new section to the end of the pywbem documentation, an
index.  Note that there are probably no hand created index entries but
the index gathers a number of entries from API information, etc.

This is parallel to change (pywbemcli pr #872) we did for pywbemcli and
only adds a new rst file (genindex.rst) and a reference to that file in
index.rst.