Issue #2490 - increase coverage
1. Increased test coverage by adding tests mostly in tests for
WBEMConnection. The tests were added in
tests/unittest/pywbem/test_cimoperationspy and in the function tests.
This added something over 100 lines tested and got us to 92%.

2. Fixed issue in WBEMConnections.enable_stats where the setter
was defined with a method rather than a property.

3. Fixed issue in WBEMConnection where open... operations inconsistently
testing OperationTimeout. Most Open... methods were testing for
None or integer but one was testing for None or positive integer.
Corrected so all test with the same method.