Fixes issue #701, SubscriptionManager bugs
1. build/use of owned object Name property.
   Fixes an issue where the build of owned CIM Listener destination and CIM
   indication filter objects and the regex to find the objects when
   restoring these objects from a server did not have a common
   understanding of the exact format of the Name property. The host_name
   element was part of the regex but not of the string in the Name
   property. This adds the host_name to the Name property.

2. build of complete instance path in _createSubsription(). Addes
   the keybindings to subscription instance in this method.

It also adds 1. test to confirm that when a SubscriptionManager is
created, the owned subscriptions, filters, and destinations are
recovered from the server correctly. 2. test that adding a
subscription a second time does not create a new subscription (i.e
adding with same paths for filter and dest as first subscription does
not create a new subscription.

This PR does not add any new subscription tests other than to
specifically test the changes.