Create pytest based test for SubscriptionManager
Adds a new group of tests for SubscriptionManager based on pytest.
This includes both api error tests and multiple test scenarios
with one or more mock wbem server, indication filters, and
indication destinations
.

Note: Test involving owned subscriptions, etc. cannot be incorporated
until issue # 2701 is completed since that affects the processing
of returned instances, etc.

This includes the ability to create the subscription manager,
multiple servers, multiple filters, and multiple destinatons.

The only validity tests to date are counts of the items created.

This pr also modifies the MockWBEMServr class to allow the use
of url in the definition of a server so that multiple servers
be defined along with multiple filters, etc.