Allow initalization defined url for FakedWBEMConnection class
Adds one more parameter to the __init__ method of FakedWBEMConnection
(url) to allow the user to define a different url that will be passed to
the superclass in place of http://FakedURL:5988.

This will be useful in testing, particularly subscription testing where
we want to establish multiple server_ids in the tests.