Add mock  user providers for filter, destination, and subscription
Adds 3 mock user providers for subscriptions, filters, and listener
destinations and tests for these new providers.

The 3 providers are in a single file since they should be initialized
together and the include a mixin class since much of the implementation
of these providers is common.

Note that this also extends the standard test_subscription tests to
integrate the user providers and to add to coverage.

The test_subscriptionmanager.py tests have been modified to use these
providers.