Fix issue #2715, SubscriptionManager loses owned parameter
The subscription manager does not pass the owned parameter to recalls of
itself if there are multiple urls provided in the call.

Extends test_suscriptions slightly

adds capability to remove destinations, filters, subscriptions and test
result in addition to new test for creating multiple destinations.
Subscription manager now up to 94% with only 15 untested lines of code.