Change use of unittest.MOCK to duck typing mocked methods in FakedWBEMConnection
The use of Mock was causing issues with pywbemcli pickle and restore of
connection information when the connections were mocked.  Replacing the
Mock with simply duck typing the methods removed the problem

See pywbemtools issueL https://github.com/pywbem/pywbemtools/issues/1038
for a detailed description of the problem.

Note that there is no test for this since pywbem does not use pickle.