Add Warning class for TolerateSchemaIssuesWarning
Adds a new warnings class to pywbem and adds that class to the unit
test.  This is expected to be used in the future in the MOF compiler and
by pywbem tools to warn the user about issues in the schema that are
either repaired or which do not warrent a failure.