Add some properties to Subscription provider instances
1. Added description property to each of the providers just to clarify
that these are from the Mock

2. Add a couple of Null properties just to expand the pywbemcli test.

3. Extended one property in subscription to add it NULL even if not
provided by client as a logical server based property.

These are all minor changes to the mock providers and have been tested
against pywbemcli.

Corrects an issue in the subscription mock providers that could cause
an exception if filter or destination instances were created through the
compiler with invalid key properties.  In one case a test was being
executed when the property did not exist or was None.