Issue #2786, compiler does not test arrayness of instance property values
The MOF compiler completely ignores issues of mismatch between the class
definition of properties arrayness and the corresponding values in
instances of the class. Thus, the compiler where a class with an array
property is defined will accept a single value property value (and
output it as a single value property value) and will accept an array
value defintion for an instance and produce an array value in the
CIMProperty.  This fix tests for this mismatch in the compiler and
generates an error if found.