Fix issue #2839, issue where DeepInheritance parameter is not passed to OpenEunumerateInstances method.
In WbemConnectionMock the DeepInheritance method is not passed to the
imeth.OpenEnumerateInstances method so the handler for this method
always returns the default DeepInheritance=True.

Adds the line of code to pass the parameter and adds the test of pull
operations for DeepInheritance to the mock tests. It also adds the
IterEnumerateInstances to the same test.