Extend associators to support bidirectional cross namespace associations.
The existing mock implementation only allowed cross namespace
associations in a single direction, from the namespace where the
association instance was created to the namespace(s) defined by a
reference property with a different namespace.

Extends CreateInstance, ModifyInstance, and DeleteInstance association
instances in pywbem_mock where the association instance reference
properties define cross namespace associations to build a corresponding
instance of the association in each namespace.  Thus, the Reference
and Association request operations will return the correct instances
when called from any namespace in which the association participates
(i.e. any namespace specified in a reference property of the
association).

The implementation creates an instance of the cross-namespace
association in each namespace in which the association participates.
All of these are exactly the same except that the namespace part
of the path is specific to each namespace.  Creating a single
cross-namespace thus creates an association instance in each namespace
in which the association participates.
Deleting the association instance deletes the association in all
namespaces in which in participates. Modifying a multi-namespace
association modifies the instance in each namespace in which the
instance participates.

This code is implemented in the _instance_write provider

ModifyInstance to the association instance  modifies the instances in all
namespaces defined and DeleteInstance deletes the instance in each
namespace.

This fixes an open issue from the original pywbem_mock.

Adds tests for all of the operations with different associations.

Adds new test file for testing pywbem_mock, test_multi_ns_assoc.py.
This file includes tests for multi-namespace associations and the
operations that use them.

Corrected new pylint found issues in several modules

Corrected some issues that apparently pylint is not finding that were
not noted before including:

1. Issue where we were using parens on return statement. Note that one
of them was valid return of a tuple but there was no space between the
return and tuple.
2. Issue where we were using parents on yield statement