Fix issue # 2865 where mocker.add_namespace fails
The failure is caused when the mock wbem connection add_namespace is
used after a namespace provider is initialized.  The new namespace is
visible from conn.namespaces but not from the WBEMServer namespace
because it was not installed using server.create_namespace().

Modifies mocker add_namespace to try to use server.create_namespace to
create the namespace and only if that fails, drop back to calling
mainprovider.add_namespace()

Adds test to confirm that add_namespace successfully creates the
namespace and that the namespace is visible from server.namespaces when
namespaces are added both before and after the namespace provider is
installed.