Fix issues with namespace provider.
This fixes two issues:

1. Error caused by bad edit in pr# 2913
2. Fix issue #2918 - limitation where duplicate namespaces could be
   created.

The fix consists of modifying the test for existence of the namespace
being created in the namespaceprovider.  A method was added to
baseprovider to accomplish this change and used in the namespace
provider.

The test modifies and extends some of the tests for namespace provider
installation