Fix mocker issue with create association instance.
Fix issue #2920 in pywbem_mock where CreateInstance or ModifyInstance on
an association instance fails if a reference property CIMInstanceName
property includes host name in the path.

Modified to fail with correct message if any reference property of an
association includes the host element in the CIMInstanceName

This seemed logical since otherwise, the user would have an instance
that they created which differed from the path we create in
CreateInstance.

Added test for this failure for CreateInstance