Fix several security issues.
Fixes safety issues for py, safety, and wheel.

Conflict in version requirements between safety (>=8.0.2) and easy-vault
(>=7.1.1) on click minimum version causes failure with 2.10 minimum because
we specify  click==7.1.1 in minimum. Moved to specify click min = 8.0.2

Conflict is caused by:
    The user requested packaging>=19.0
    pytest 6.2.5 depends on packaging
    build 0.5.1 depends on packaging>=19.0
    safety 2.2.0 depends on packaging>=21.0
    The user requested (constraint) packaging==19.0

So new constraint is packaging 21.0 if safety 2.2.0

NOTE: These are the same fixes made for pywbemtools in November.