Fix issue with possible HTTP retries of cim-xml operations.
Fixes an issue in the parameters for  HTTP POST retries within
urllib3 where the retry values defined in CIM_operations  allowed
the retry of a CIM/XML request that such as
invokemethod, update, etc. where it is critical that the operation only
be issued once.  This is because all CIM/XML operations use POST.

Changes the retry to not allow retries on POST operations.

Issue documented in issue #2951