Changes to indications example
Add Code for display of request/response log

Fix issues in the example pegasusindicationtest.py.

This example was written against earlier version of pywbem and
the interfaces with subscription manager have changed.

Also cleaned up some of the code and updated print statements.

Finally extended the test so that it could be executed against a
pegasus WBEM server in a container.