Rewrite the example pegasusindicationtest.py to be usable in a container environment and also to clean up the code.
This included changing the arg processor to use argparse and
rewriting much of the code to make it cleaner and to output more
meaningful messages.

NOTE: This code is not run as part of our tests.  Also, there
still appears to be a problem but it is apparenntly with the container
and not Pegasus itself in that often the listener only receives
a subset of the indicaitons and the remainder are still queued for
output in OpenPegasus (our displays show that when the test fails)