Expand documentation on the WBEMListener host parameter
This change clarifies the definition of the host parameter on the
WBEMListener class to include the definition of how to define a
listener that listens on multiple network addresses.

Note that there is still some ambiguity in that a. this does not
discusss IPV6 and b) only briefly mentions that the actual behavior
of a listener with a specific address may be OS or network dependent in
that in some cases the test is limited to only pass requests with
the specific address and in other cases, it will pass any address that
is part of the specific network interface defined by the host parameter.