Add end2end test for indications from OpenPegasus in container
This test is based on the pegasusindiationtest.py example in
the examples directory.

NOTE: We modified the example code somewhat from what was already
committed to have the same code for the test class itself to
handle differences in the setup with the end2end test and to keep
the majority of the code aligned.