Update documentation on listener host parameter. (#3059)
* Clarify documentation of listener host parameter

Clarifies the definition of this parameter, in particular the wildcard
IP addresses and the alternative empty string to represent the wild
card.

* Modify example of setting listener bind address

1. Clarify the explanation of what values can be supplied for the bind
address to the listener (i.e. the host parameter).

2. Modify the example because getfqdn() may return localhost which is
useful only if the the indication sender is on the same system.  This
also shows the concept of the wildcard bind address which is the
standard example used by python in examples of setting up web listeners,
etc.

---------

Co-authored-by: kschopmeyer <k.schopmeyerwork@gmail.com>