Add documentation in the update of urllib3
The inclusion of urllib3 verson 2.0+ can cause SSL incompatibility
issue because it sets the minimum TLS version to 1.2 where
it was previously TLS 1.1 or less and also limits the
allowable SSL libraries to OpenSSL and possibly libreSSL.

This change updates the changes.rst document to describe the
possible issues and also updates the troubleshooting section
to provide details on the possible issues and corrections.

Adds index entries in a number of places.