Eliminate invalid WARNING: reference not found msgs from sphinx.
See issue #3064 for more detailed information on the warning messages.
The full list of warnings output is included in issue #3064

This PR fixes the issues by:

* Disabling  about 270 of the issues  using the nitpick-ignore parameter
in the Sphinx conf.py file.  These appear to be issues with finding
links of methods in  superclasses.

* Fixing several warning that were issues with formatting the
  docstrings. In general the issues were things like missing the
  ~pywbem component of target reference, invalid indent on second
  line of parameter defintion, invalid object reference on
  Raises: definition, etc.

* One issue  __logging where the code itself was fixed by adding
  a module global to the __all__ list.

Currently there are now no Warning invalid reference messages appearing
on a full make builddoc.