Fix issue in end2end/test_indications
Fix issue documented in issue #3071.  This was a code error. There was a
case where the time for the last indication received variable was not
set correctly.

This also adds the last sequence number to the error and changes  from a
display of the issue to an assert. The test should fail if the
indications were not received as a monotonically increasing sequence
with no gaps in the sequence.