Fix issue #3071, end2end indication test failure.
Corrected code issue where the last indication time was not always
captured before testing for elapsed time.

Also modified the code testing for next sequence number to assert if it
is not correct since that is actually an error.

Not in changes since this is a change within a single pywbem
version