Fixes issue #3071 end end test fail. (#3076)
* Fix issue in end2end/test_indications

Fix issue documented in issue #3071.  This was a code error. There was a
case where the time for the last indication received variable was not
set correctly.

This also adds the last sequence number to the error and changes  from a
display of the issue to an assert. The test should fail if the
indications were not received as a monotonically increasing sequence
with no gaps in the sequence.

* Fix issue #3071, end2end indication test failure.

Corrected code issue where the last indication time was not always
captured before testing for elapsed time.

Also modified the code testing for next sequence number to assert if it
is not correct since that is actually an error.

Not in changes since this is a change within a single pywbem
version

---------

Co-authored-by: kschopmeyer <k.schopmeyerwork@gmail.com>