Modify listener to provide interprocess queue.
Modify pywbem listener to provide for interprocess queue between
listener and callback.  The listener puts  received indications into
the python interprocess queue and the callback retreives
indications from that queue.

Currently the listener code includes a flag so that this queue can be
enabled or disabledi (i.e. implementing the original behavior and the
queue).  Note that enabling this queue tends to increase
the number of indication failures somewhat.

We have NOT yet identified the issue causing the indication failures that
occur frequently in pywbemcli but infrequently in pywbem itself.

This change also modifies the examples/pegasus indication test somewhat
to allow separating the parts of the test and to allow the test code to
determine the listener destination host IP as an option. These code
changes in the example help with testing the indication failure.